+++ python_modules/dagit/dagit_tests/job.py
@@ -1,27 +1,23 @@
-from dagster import Int, repository
+from dagster import job, In, Out, op, Int, repository
 from dagster._core.test_utils import today_at_midnight
 from dagster._legacy import (
-    InputDefinition,
-    OutputDefinition,
     daily_schedule,
-    lambda_solid,
-    pipeline,
 )
 
 
-@lambda_solid(input_defs=[InputDefinition("num", Int)], output_def=OutputDefinition(Int))
+@op(ins={"num": In(Int)}, out=Out(Int))
 def add_one(num):
     return num + 1
 
 
-@lambda_solid(input_defs=[InputDefinition("num", Int)], output_def=OutputDefinition(Int))
+@op(ins={"num": In(Int)}, out=Out(Int))
 def mult_two(num):
     return num * 2
 
 
-@pipeline
+@job
 def math():
-    return mult_two(num=add_one())
+    mult_two(num=add_one())
 
 
 @daily_schedule(
