+++ python_modules/dagit/dagit_tests/override.yaml
@@ -1,4 +1,4 @@
 load_from:
   - python_file:
-      relative_path: override_pipeline.py
+      relative_path: override_job.py
       location_name: load_from_file
