+++ python_modules/dagit/dagit_tests/telemetry_repository.yaml
@@ -3,5 +3,5 @@ load_from:
       relative_path: ../../dagster/dagster_tests/general_tests/test_repository.py
       attribute: dagster_test_repository
   - python_file:
-      relative_path: pipeline.py
+      relative_path: job.py
       attribute: test_repository
