+++ python_modules/dagit/dagit_tests/test_subscriptions.py
@@ -12,8 +12,8 @@ from starlette.testclient import TestClient
 from dagster._core.test_utils import environ, instance_for_test
 from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster._core.workspace.load_target import WorkspaceFileTarget
-from dagster._legacy import execute_pipeline, pipeline, solid
 from dagster._utils import file_relative_path
+from dagster import job, op
 
 EVENT_LOG_SUBSCRIPTION = """
     subscription PipelineRunLogsSubscription($runId: ID!) {
@@ -73,19 +73,19 @@ def end_subscription(ws):
     ws.close()
 
 
-@solid
-def example_solid():
+@op
+def example_op():
     return 1
 
 
-@pipeline
-def example_pipeline():
-    example_solid()
+@job
+def example_job():
+    example_op()
 
 
 def test_event_log_subscription():
     with instance_for_test() as instance:
-        run = execute_pipeline(example_pipeline, instance=instance)
+        run = example_job.execute_in_process(instance=instance)
         assert run.success
         assert run.run_id
 
@@ -108,7 +108,7 @@ def test_event_log_subscription():
 )
 def test_event_log_subscription_chunked():
     with instance_for_test() as instance, environ({"DAGIT_EVENT_LOAD_CHUNK_SIZE": "2"}):
-        run = execute_pipeline(example_pipeline, instance=instance)
+        run = example_job.execute_in_process(instance=instance)
         assert run.success
         assert run.run_id
 
@@ -133,7 +133,7 @@ def test_compute_log_subscription(mock_watch_completed):
     mock_watch_completed.return_value = False
 
     with instance_for_test() as instance:
-        run = execute_pipeline(example_pipeline, instance=instance)
+        run = example_job.execute_in_process(instance=instance)
         assert run.success
         assert run.run_id
 
@@ -146,7 +146,7 @@ def test_compute_log_subscription(mock_watch_completed):
                     COMPUTE_LOG_SUBSCRIPTION,
                     {
                         "runId": run.run_id,
-                        "stepKey": "example_solid",
+                        "stepKey": "example_op",
                         "ioType": "STDERR",
                     },
                 )
