+++ python_modules/dagit/dagit_tests/workspace.yaml
@@ -1,4 +1,4 @@
 load_from:
   - python_file:
-      relative_path: pipeline.py
+      relative_path: job.py
       location_name: load_from_file
